Tighten browser manager extraction and CI linting
